requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==3.2.25
django-cacheops==7.1
funcy==2.0
idna==3.10
legacy-cgi==2.6.3
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
